Refactor to make fragment module less unwieldy
